+++ examples/vanilla-spa-i18n-baked/src/main.ts
@@ -25,7 +25,7 @@ document.querySelector<HTMLDivElement>("#app")!.innerHTML = `
   </div>
   <div>
     <h1>Get started</h1>
-    <p>Edit <!-- @zintl-ignore --> <code>src/main.ts</code> and save to test <!-- @zintl-ignore --> <code>HMR</code></p>
+    <p>Edit <code>src/main.ts</code> and save to test <code>HMR</code></p>
   </div>
   <button id="counter" type="button" class="counter"></button>
 </section>